Fixed home screen preview initial state
