+++ unstoppable-ios-app/domains-manager-ios-preview/Modules/PreviewDomainsCollection/PreviewDomainsCollectionViewController.swift
@@ -16,3 +16,82 @@ import SwiftUI
     
     return nav
 }
+
+@MainActor
+final class PreviewDomainsCollectionViewPresenter {
+    private weak var view: DomainsCollectionViewProtocol?
+    var analyticsName: Analytics.ViewName {
+        .home
+    }
+    
+    var navBackStyle: BaseViewController.NavBackIconStyle {
+        .arrow
+    }
+    
+    init(view: DomainsCollectionViewProtocol) {
+        self.view = view
+    }
+}
+
+// MARK: - DomainsCollectionPresenterProtocol
+extension PreviewDomainsCollectionViewPresenter: DomainsCollectionPresenterProtocol {
+    func viewDidLoad() {
+        UserDefaults.didShowSwipeDomainCardTutorial = false
+        view?.setGoToSettingsTutorialHidden(true) // Always hide for now (MOB-394)
+        view?.setScanButtonHidden(true)
+        view?.setAddButtonHidden(false, isMessagingAvailable: false)
+        view?.setEmptyState(hidden: true)
+        view?.setSelectedDisplayMode(.empty, at: 0, animated: false)
+//        view?.setSelectedDisplayMode(.domain(.init(name: "oleg.x", ownerWallet: "", isSetForRR: false)), at: 0, animated: false)
+    }
+    
+    var currentIndex: Int {
+        0
+    }
+    
+    func canMove(to index: Int) -> Bool {
+        false
+    }
+    
+    func displayMode(at index: Int) -> DomainsCollectionCarouselItemDisplayMode? {
+        .empty
+    }
+    
+    func didMove(to index: Int) {
+        
+    }
+    
+    func didOccureUIAction(_ action: DomainsCollectionViewController.Action) {
+        
+    }
+    
+    func didTapSettingsButton() {
+        
+    }
+    
+    func importDomainsFromWebPressed() {
+        
+    }
+    
+    func didPressScanButton() {
+        
+    }
+    
+    func didMintDomains(result: MintDomainsResult) {
+        
+    }
+    
+    func didRecognizeQRCode() {
+        
+    }
+    
+    func didTapAddButton() {
+        
+    }
+    
+    func didTapMessagingButton() {
+        
+    }
+    
+    
+}
